move validateSupplier method to service layer
