+++ src/main/java/com/mercuryCyclists/Inventory/service/ProductService.java
@@ -5,6 +5,7 @@ import com.mercuryCyclists.Inventory.entity.Product;
 import com.mercuryCyclists.Inventory.repository.PartRepository;
 import com.mercuryCyclists.Inventory.repository.ProductRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.List;
 import java.util.Set;
@@ -13,14 +14,17 @@ import java.util.Set;
 public class ProductService {
     private ProductRepository productRepository;
     private PartRepository partRepository;
+    private RestfulService restfulService;
 
-    public ProductService(ProductRepository productRepository, PartRepository partRepository) {
+    public ProductService(ProductRepository productRepository, PartRepository partRepository, RestfulService restfulService) {
         this.productRepository = productRepository;
         this.partRepository = partRepository;
+        this.restfulService = restfulService;
     }
 
     /**
      * Add a product to repo
+     *
      * @param product
      * @return a product
      */
@@ -30,6 +34,7 @@ public class ProductService {
 
     /**
      * Get all products from the repo
+     *
      * @return all products
      */
     public List<Product> getAllProducts() {
@@ -38,6 +43,7 @@ public class ProductService {
 
     /**
      * Get a product according to id
+     *
      * @param id
      * @return a product with the given id
      */
@@ -47,6 +53,7 @@ public class ProductService {
 
     /**
      * Update the details of a product from repo
+     *
      * @param product
      * @param id
      * @return product with updated details
@@ -64,6 +71,7 @@ public class ProductService {
 
     /**
      * Delete a product with given id from repo
+     *
      * @param id
      */
     public void deleteProduct(Long id) {
@@ -72,21 +80,39 @@ public class ProductService {
     }
 
     /**
+     * Helper method, check if a supplier id is valid
+     *
+     * @param supplierId
+     * @return true if id is valid else false
+     */
+    private boolean validSupplier(Long supplierId) {
+        try {
+            restfulService.getSupplier(supplierId).getStatusCodeValue();
+            return true;
+        } catch (HttpServerErrorException e) {
+            return false;
+        }
+    }
+
+    /**
      * Helper method
+     *
      * @param id
      * @return product if product exist else null
      */
     private Product productExist(Long id) {
         return productRepository.findById(id).orElse(null);
     }
+
     /**
      * Add a part to repo
+     *
      * @param part
      * @return a part
      */
     public boolean addPart(Long productId, Part part) {
         Product p = productExist(productId);
-        if (p == null) return false;
+        if (p == null || !validSupplier(part.getSupplierId())) return false;
         // add part to product set
         p.getSet().add(part);
         updateProduct(p, productId);
@@ -95,6 +121,7 @@ public class ProductService {
 
     /**
      * Update part helper method
+     *
      * @param product
      * @param part
      * @return If given product contains given part return true else false
@@ -105,6 +132,7 @@ public class ProductService {
 
     /**
      * Update a part with given ID
+     *
      * @param productId
      * @param partId
      * @param part
@@ -113,14 +141,17 @@ public class ProductService {
     public Part updatePart(Long productId, Long partId, Part part) {
         Product product = productRepository.findById(productId).orElse(null);
         Part p = partRepository.findById(partId).orElse(null);
-        if (p == null || product == null || !updatePartHelper(product, p)) return null;
+        if (p == null || product == null || !updatePartHelper(product, p) || !validSupplier(part.getSupplierId()))
+            return null;
         p.setName(part.getName());
         p.setDescription(part.getDescription());
+        p.setQuantity(part.getQuantity());
         return partRepository.save(p);
     }
 
     /**
      * Get all the parts with given product id
+     *
      * @param id
      * @return all parts for a product
      */
@@ -132,6 +163,7 @@ public class ProductService {
 
     /**
      * Delete a part with given product id and part id
+     *
      * @param productId
      * @param partId
      */
